Rename Cardano.CLI.Shelley.Key to Cardano.CLI.Types.Key
