+++ cardano-cli/src/Cardano/CLI/Commands/Legacy.hs
@@ -55,7 +55,7 @@ import           Cardano.Api.Shelley hiding (QueryInShelleyBasedEra (..))
 import           Cardano.Chain.Common (BlockCount)
 import           Cardano.CLI.Types.Governance
 import           Cardano.CLI.Options.Governance
-import           Cardano.CLI.Shelley.Key (DelegationTarget, PaymentVerifier, StakeIdentifier,
+import           Cardano.CLI.Types.Key (DelegationTarget, PaymentVerifier, StakeIdentifier,
                    StakeVerifier, VerificationKeyOrFile, VerificationKeyOrHashOrFile,
                    VerificationKeyTextOrFile)
 import           Cardano.CLI.Types.Legacy
