+++ cardano-cli/src/Cardano/CLI/Options/Governance.hs
@@ -8,8 +8,8 @@ import           Cardano.Api.Shelley
 
 import           Cardano.CLI.Environment
 import           Cardano.CLI.Options.Common
-import           Cardano.CLI.Shelley.Key
 import           Cardano.CLI.Types.Governance
+import           Cardano.CLI.Types.Key
 import           Cardano.CLI.Types.Legacy
 import           Cardano.Ledger.Shelley.TxBody (MIRPot)
