+++ cardano-cli/src/Cardano/CLI/Run/Legacy/Address.hs
@@ -20,7 +20,7 @@ import           Cardano.Api.Shelley
 import           Cardano.CLI.Options.Legacy (AddressCmd (..), AddressKeyType (..))
 import           Cardano.CLI.Run.Legacy.Address.Info (ShelleyAddressInfoError, runAddressInfo)
 import           Cardano.CLI.Run.Legacy.Read
-import           Cardano.CLI.Shelley.Key (PaymentVerifier (..), StakeIdentifier (..),
+import           Cardano.CLI.Types.Key (PaymentVerifier (..), StakeIdentifier (..),
                    StakeVerifier (..), VerificationKeyTextOrFile,
                    VerificationKeyTextOrFileError (..), generateKeyPair, readVerificationKeyOrFile,
                    readVerificationKeyTextOrFileAnyOf, renderVerificationKeyTextOrFileError)
