+++ cardano-cli/src/Cardano/CLI/Run/Legacy/Key.hs
@@ -21,7 +21,7 @@ import           Cardano.Api.Shelley
 
 import qualified Cardano.CLI.Byron.Key as Byron
 import           Cardano.CLI.Commands.Legacy
-import           Cardano.CLI.Shelley.Key (VerificationKeyTextOrFile (..),
+import           Cardano.CLI.Types.Key (VerificationKeyTextOrFile (..),
                    VerificationKeyTextOrFileError, readVerificationKeyTextOrFileAnyOf,
                    renderVerificationKeyTextOrFileError)
 import           Cardano.CLI.Types.Legacy (SigningKeyFile, VerificationKeyFile)
