+++ cardano-cli/src/Cardano/CLI/Run/Legacy/Node.hs
@@ -15,7 +15,7 @@ import           Cardano.Api
 import           Cardano.Api.Shelley
 
 import           Cardano.CLI.Commands.Legacy
-import           Cardano.CLI.Shelley.Key (VerificationKeyOrFile, readVerificationKeyOrFile)
+import           Cardano.CLI.Types.Key (VerificationKeyOrFile, readVerificationKeyOrFile)
 import           Cardano.CLI.Types.Legacy (KeyOutputFormat (..), SigningKeyFile, VerificationKeyFile)
 
 import           Control.Monad.IO.Class (MonadIO (..))
