+++ cardano-cli/src/Cardano/CLI/Run/Legacy/Pool.hs
@@ -12,7 +12,7 @@ import           Cardano.Api
 import           Cardano.Api.Shelley
 
 import           Cardano.CLI.Commands.Legacy
-import           Cardano.CLI.Shelley.Key (VerificationKeyOrFile, readVerificationKeyOrFile)
+import           Cardano.CLI.Types.Key (VerificationKeyOrFile, readVerificationKeyOrFile)
 import           Cardano.CLI.Types.Legacy (PoolIdOutputFormat (..))
 import qualified Cardano.Ledger.Slot as Shelley
